requirements.txt
Django==1.6
psycopg2==2.4.5
dj-database-url==0.2.0
South
django-tastypie
